rectification put adr

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: lmatkows <lmatkows@student.42perpignan.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/31 10:14:27 by Lmatkows          #+#    #+#             */
-/*   Updated: 2024/11/12 09:30:49 by lmatkows         ###   ########.fr       */
+/*   Updated: 2024/11/12 09:49:49 by lmatkows         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,8 +20,8 @@ int		ft_putchar(int c);
 int		ft_putstr(char *s);
 int		ft_putdec(int d);
 int		ft_putundec(unsigned int u);
-int		ft_putlowhexa(unsigned int x);
-int		ft_putuphexa(unsigned int X);
+int		ft_putlowhexa(unsigned long x);
+int		ft_putuphexa(unsigned long X);
 int		ft_putadrhexa(void *p);
 int		ft_findset(char c, char *set);
 char	ft_symb(const char *arg, int i);
